Create ListChoicePrompt
